feat(settings): Move settings from the modal into a tab panel.
