+++ src/components/CentralContainer/Sidebar/SidebarTabs/SettingsTabPanel/index.css
@@ -1,0 +1,21 @@
+.settings-tab-container {
+    overflow: hidden;
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+}
+
+.settings-form {
+    overflow-y: hidden;
+    display: flex;
+    flex-direction: column;
+    gap: 0.75rem;
+}
+
+.settings-form-fields-container {
+    overflow-y: auto;
+    display: flex;
+    flex-direction: column;
+    flex-grow: 1;
+    gap: 0.75rem;
+}
