--- src/components/modals/SettingsModal/index.css
@@ -1,12 +1,0 @@
-.settings-dialog-title {
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    justify-content: space-between;
-
-    font-size: 2rem !important;
-}
-
-.config-form-control {
-    margin-top: 16px;
-}
